Scala enumeration works strangely. Hopefully this fixes issue #7.
The issue was that invoking toString on values in Enumeration caused the
Enumeration to try to use reflection in quite a weird manner. It first
searched for method with given signature and later tried to invoke this
method. Apparently they found method that behaved differently than
expected and this resulted in

AssertionError: assertion failed: Duplicate id: 12

I have tried to reproduce this problem on an isolated example, but
meanwhile all the enumeration objects just get a string description.